+++ injectable_generator/lib/code_builder/allocator.dart
@@ -24,5 +24,6 @@ class HashedAllocator implements Allocator {
   int _hashedUrl() => _url.hashCode / 1000000 ~/ 1;
 
   @override
-  Iterable<Directive> get imports => _imports.keys.map((u) => Directive.import(u, as: '_i${_imports[u]}'));
+  Iterable<Directive> get imports =>
+      _imports.keys.map((u) => Directive.import(u, as: '_i${_imports[u]}'));
 }