+++ injectable_generator/lib/lean_builder/lean_injectable_builder.dart
@@ -62,10 +62,13 @@ class LeanInjectableBuilder extends Builder {
         for (var element in executables) {
           if (element.isPrivate) continue;
           allDepsInStep.add(
-            LeanDependencyResolver(resolver).resolveModuleMember(clazz, element),
+            LeanDependencyResolver(
+              resolver,
+            ).resolveModuleMember(clazz, element),
           );
         }
-      } else if (_hasInjectable(clazz) || (_autoRegister && _hasConventionalMatch(clazz))) {
+      } else if (_hasInjectable(clazz) ||
+          (_autoRegister && _hasConventionalMatch(clazz))) {
         allDepsInStep.add(
           LeanDependencyResolver(resolver).resolve(clazz),
         );
@@ -91,8 +94,12 @@ class LeanInjectableBuilder extends Builder {
     if (clazz.hasAbstract) {
       return false;
     }
-    final fileName = clazz.thisType.declarationRef.srcUri.path.replaceFirst('.dart', '');
-    return (_classNameMatcher != null && _classNameMatcher!.hasMatch(clazz.name)) ||
+    final fileName = clazz.thisType.declarationRef.srcUri.path.replaceFirst(
+      '.dart',
+      '',
+    );
+    return (_classNameMatcher != null &&
+            _classNameMatcher!.hasMatch(clazz.name)) ||
         (_fileNameMatcher != null && _fileNameMatcher!.hasMatch(fileName));
   }
 }