+++ injectable_generator/lib/lean_builder/resolvers/lean_dependency_resolver.dart
@@ -12,19 +12,46 @@ import 'package:lean_builder/type.dart';
 
 import '../build_utils.dart';
 
-const _injectableChecker = TypeChecker.typeNamed(Injectable, inPackage: 'injectable');
-const _lazySingletonChecker = TypeChecker.typeNamed(LazySingleton, inPackage: 'injectable');
-const _singletonChecker = TypeChecker.typeNamed(Singleton, inPackage: 'injectable');
+const _injectableChecker = TypeChecker.typeNamed(
+  Injectable,
+  inPackage: 'injectable',
+);
+const _lazySingletonChecker = TypeChecker.typeNamed(
+  LazySingleton,
+  inPackage: 'injectable',
+);
+const _singletonChecker = TypeChecker.typeNamed(
+  Singleton,
+  inPackage: 'injectable',
+);
 
 const _namedChecker = TypeChecker.typeNamed(Named, inPackage: 'injectable');
-const _ignoredChecker = TypeChecker.typeNamed(IgnoreParam, inPackage: 'injectable');
+const _ignoredChecker = TypeChecker.typeNamed(
+  IgnoreParam,
+  inPackage: 'injectable',
+);
 const _envChecker = TypeChecker.typeNamed(Environment, inPackage: 'injectable');
-const _preResolveChecker = TypeChecker.typeNamed(PreResolve, inPackage: 'injectable');
-const _factoryParamChecker = TypeChecker.typeNamed(FactoryParam, inPackage: 'injectable');
+const _preResolveChecker = TypeChecker.typeNamed(
+  PreResolve,
+  inPackage: 'injectable',
+);
+const _factoryParamChecker = TypeChecker.typeNamed(
+  FactoryParam,
+  inPackage: 'injectable',
+);
 const _scopeChecker = TypeChecker.typeNamed(Scope, inPackage: 'injectable');
-const _factoryMethodChecker = TypeChecker.typeNamed(FactoryMethod, inPackage: 'injectable');
-const _disposeMethodChecker = TypeChecker.typeNamed(DisposeMethod, inPackage: 'injectable');
-const _postConstructChecker = TypeChecker.typeNamed(PostConstruct, inPackage: 'injectable');
+const _factoryMethodChecker = TypeChecker.typeNamed(
+  FactoryMethod,
+  inPackage: 'injectable',
+);
+const _disposeMethodChecker = TypeChecker.typeNamed(
+  DisposeMethod,
+  inPackage: 'injectable',
+);
+const _postConstructChecker = TypeChecker.typeNamed(
+  PostConstruct,
+  inPackage: 'injectable',
+);
 const _orderChecker = TypeChecker.typeNamed(Order, inPackage: 'injectable');
 
 class LeanDependencyResolver {
@@ -105,22 +132,29 @@ class LeanDependencyResolver {
   ]) {
     final annotatedElement = excModuleMember ?? clazz;
     _typeImpl = _type;
-    var injectableAnnotation = _injectableChecker.firstAnnotationOf(annotatedElement);
+    var injectableAnnotation = _injectableChecker.firstAnnotationOf(
+      annotatedElement,
+    );
     DartType? abstractType;
     ExecutableElement? disposeFuncFromAnnotation;
     List<String>? inlineEnv;
-    if (injectableAnnotation != null && injectableAnnotation.constant is ConstObject) {
+    if (injectableAnnotation != null &&
+        injectableAnnotation.constant is ConstObject) {
       final injectable = injectableAnnotation.constant as ConstObject;
 
       _cache = injectable.getBool('cache')?.value == true;
 
       if (_lazySingletonChecker.isExactlyType(injectable.type)) {
         _injectableType = InjectableType.lazySingleton;
-        disposeFuncFromAnnotation = injectable.getFunctionReference('dispose')?.element;
+        disposeFuncFromAnnotation = injectable
+            .getFunctionReference('dispose')
+            ?.element;
       } else if (_singletonChecker.isExactlyType(injectable.type)) {
         _injectableType = InjectableType.singleton;
         _signalsReady = injectable.getBool('signalsReady')?.value;
-        disposeFuncFromAnnotation = injectable.getFunctionReference('dispose')?.element;
+        disposeFuncFromAnnotation = injectable
+            .getFunctionReference('dispose')
+            ?.element;
         var dependsOn = injectable
             .getList('dependsOn')
             ?.value
@@ -134,7 +168,11 @@ class LeanDependencyResolver {
         }
       }
       abstractType = injectable.getTypeRef('as')?.value;
-      inlineEnv = injectable.getList('env')?.literalValue.cast<String>().toList();
+      inlineEnv = injectable
+          .getList('env')
+          ?.literalValue
+          .cast<String>()
+          .toList();
       _scope = injectable.getString('scope')?.value;
       _order = injectable.getInt('order')?.value;
     }
@@ -167,17 +205,26 @@ class LeanDependencyResolver {
             .map<String>((e) => e.getString('name')!.value)
             .toList();
 
-    if (_scopeChecker.firstAnnotationOfExact(annotatedElement) case var scopeAnnotation?) {
-      _scope ??= (scopeAnnotation.constant as ConstObject).getString('name')?.value;
+    if (_scopeChecker.firstAnnotationOfExact(annotatedElement)
+        case var scopeAnnotation?) {
+      _scope ??= (scopeAnnotation.constant as ConstObject)
+          .getString('name')
+          ?.value;
     }
 
     _preResolve = _preResolveChecker.hasAnnotationOfExact(annotatedElement);
-    if (_orderChecker.firstAnnotationOfExact(annotatedElement) case var orderAnnotation?) {
-      _order = (orderAnnotation.constant as ConstObject).getInt('position')?.value;
+    if (_orderChecker.firstAnnotationOfExact(annotatedElement)
+        case var orderAnnotation?) {
+      _order = (orderAnnotation.constant as ConstObject)
+          .getInt('position')
+          ?.value;
     }
 
-    if (_namedChecker.firstAnnotationOfExact(annotatedElement) case var nameAnnotation?) {
-      final nameValue = (nameAnnotation.constant as ConstObject).getString('name')?.value;
+    if (_namedChecker.firstAnnotationOfExact(annotatedElement)
+        case var nameAnnotation?) {
+      final nameValue = (nameAnnotation.constant as ConstObject)
+          .getString('name')
+          ?.value;
       if (nameValue is String) {
         if (nameValue.isNotEmpty) {
           _instanceName = nameValue;
@@ -209,7 +256,8 @@ class LeanDependencyResolver {
     } else if (disposeFuncFromAnnotation != null) {
       final params = disposeFuncFromAnnotation.parameters;
       throwIf(
-        params.length != 1 || _typeResolver.resolveType(params.first.type) != _type,
+        params.length != 1 ||
+            _typeResolver.resolveType(params.first.type) != _type,
         'Dispose function for $_type must have the same signature as FutureOr Function($_type instance)',
         element: disposeFuncFromAnnotation,
       );
@@ -233,7 +281,9 @@ class LeanDependencyResolver {
 
       executableInitializer = possibleFactories.firstWhere(
         (m) {
-          final annotation = _factoryMethodChecker.firstAnnotationOf(m)?.constant;
+          final annotation = _factoryMethodChecker
+              .firstAnnotationOf(m)
+              ?.constant;
           if (annotation is ConstObject) {
             _preResolve |= annotation.getBool('preResolve')?.value ?? false;
             return true;
@@ -281,8 +331,11 @@ class LeanDependencyResolver {
         );
         continue;
       }
-      final namedAnnotation = _namedChecker.firstAnnotationOf(param)?.constant as ConstObject?;
-      final instanceName = namedAnnotation?.getTypeRef('type')?.value.name ?? namedAnnotation?.getString('name')?.value;
+      final namedAnnotation =
+          _namedChecker.firstAnnotationOf(param)?.constant as ConstObject?;
+      final instanceName =
+          namedAnnotation?.getTypeRef('type')?.value.name ??
+          namedAnnotation?.getString('name')?.value;
 
       final resolvedType = param.type is FunctionType
           ? _typeResolver.resolveFunctionType(param.type as FunctionType)
@@ -308,8 +361,12 @@ class LeanDependencyResolver {
     }
 
     _canBeConst =
-        (executableInitializer is ConstructorElement && executableInitializer.isConst) && _dependencies.isEmpty;
-    final factoryParamsCount = _dependencies.where((d) => d.isFactoryParam).length;
+        (executableInitializer is ConstructorElement &&
+            executableInitializer.isConst) &&
+        _dependencies.isEmpty;
+    final factoryParamsCount = _dependencies
+        .where((d) => d.isFactoryParam)
+        .length;
 
     throwIf(
       _preResolve && factoryParamsCount != 0,
@@ -374,10 +431,13 @@ class LeanDependencyResolver {
         _isAsync = method.returnType.isDartAsyncFuture;
         _preResolve = postConstructAnnotation.getBool('preResolve')!.value;
         final returnType = _typeResolver.resolveType(
-          _isAsync ? (method.returnType as ParameterizedType).typeArguments.first : method.returnType,
+          _isAsync
+              ? (method.returnType as ParameterizedType).typeArguments.first
+              : method.returnType,
         );
 
-        postConstructReturnsSelf = returnType == _type || returnType == _typeImpl;
+        postConstructReturnsSelf =
+            returnType == _type || returnType == _typeImpl;
         break;
       }
     }