+++ injectable/CHANGELOG.md
@@ -1,4 +1,6 @@
 # ChangeLog
+## [2.7.1+3]
+- Chore: update docs to include lean_builder support
 ## [2.7.1+2]
 - Fix: fix coverage badge alignment in readme
 ## [2.7.1+1]